app/requirements.txt
streamlit>=1.36
google-genai>=0.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
numpy>=1.26.0
plotly>=5.0.0
torchvision>=0.18.0
